Instagram_killer/pyproject.toml
[tool.poetry]
name = "killer-instagram"
version = "0.1.0"
description = ""
authors = ["KossKokos <koskokos0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10"
alembic = "1.10.2"
babel = "2.13.0"
bcrypt = "4.0.1"
cloudinary = "1.32.0"
colorama = "0.4.6"
cryptography = "41.0.5"
docutils = "0.19"
fastapi = "0.95.0"
email-validator = "1.3.1"
fastapi-limiter = "0.1.5"
fastapi-mail = "1.2.7"
httpx = "0.25.2"
jinja2 = "3.1.2"
passlib = "1.7.4"
psycopg2 = "2.9.5"
pydantic = "1.10.7"
pytest = "7.4.3"
pytest-mock = "3.12.0"
python-dotenv = "1.0.0"
python-jose = "3.3.0"
python-multipart = "0.0.6"
redis = "4.5.4"
requests = "2.31.0"
sniffio = "1.3.0"
sphinx = "6.1.3"
sqlalchemy = "2.0.7"
starlette = "0.26.1"
urllib3 = "1.26.18"
uvicorn = "0.21.1"
pytest-asyncio = "0.23.2"
pytest-trio = "0.8.0"
qrcode = {version = "7.4.2", extras = ["pil"]}


[tool.poetry.group.dev.dependencies]
sphinx = "6.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]